Move CLI for silence-lint-errors into its own package
