+++ setup.cfg
@@ -25,7 +25,7 @@ python_requires = >=3.9
 
 [options.entry_points]
 console_scripts =
-    silence-lint-error = silence_lint_error.silence_lint_error:main
+    silence-lint-error = silence_lint_error.cli.silence_lint_error:main
     fix-silenced-error = silence_lint_error.fix_silenced_error:main
 
 [bdist_wheel]
